+++ oskill/data/__init__.py
@@ -6,6 +6,8 @@ from datetime import timedelta
 
 import pandas as pd
 
+from oskill.data.calendar_surprise_detect import calendar_surprise_detect
+
 
 def point_in_time_join(
     *,